CIC2 data check tool modified

+++ HWInterface/D19cFWInterface.h
@@ -397,7 +397,7 @@ namespace Ph2_HwInterface {
 
         // CDCE 
         void configureCDCE_old(uint16_t pClockRate=120 );
-        void configureCDCE( uint16_t pClockRate=120, std::pair<std::string,uint16_t> pCDCEselect=std::make_pair("sec",40) ); 
+        void configureCDCE( uint16_t pClockRate=120, std::pair<std::string,float> pCDCEselect=std::make_pair("sec",40) ); 
         void syncCDCE();
         void epromCDCE();
